fix(from-event): return the subject and track change of element.
Set is called more than one time, and it could be called with `undefined` (when no dom element matches). On every set, the element should be new and a new `fromEvent` must be created.

We init if subject doesn't exist yet, check if there's a subscription and unsubscribe from it, check if elementRef isn't `undefined` and remove event from future getters (because that `event` doesn't exists now). Once, we reach the creation of the event, create it and subscribe to it using the shared function `subscribeToEventIfPossible` (using the setter, this function always should create a subscription).